Problema 1

+++ Logic/adunare_valori.py
@@ -1,7 +1,7 @@
 from Domain.cheltuiala import creeaza_cheltuiala, get_data, get_id, get_numar_apartament, get_suma, get_tip
 
 
-def adunare_valoare(lst_cheltuieli, data, val:int, undo_list, redo_list):
+def adunare_valoare(lst_cheltuieli, data, val:float, undo_list, redo_list):
     '''
     Adunarea unei valori tuturor cheltuilelior dintr-o data citita
     param: lst_cheltuieli - lista cheltuielilor
@@ -18,7 +18,7 @@ def adunare_valoare(lst_cheltuieli, data, val:int, undo_list, redo_list):
     
     for cheltuiala in lst_cheltuieli:
         if data == get_data(cheltuiala):
-            suma_noua = val + int(get_suma(cheltuiala))
+            suma_noua = val + float(get_suma(cheltuiala))
             result.append(creeaza_cheltuiala(get_id(cheltuiala), get_numar_apartament(cheltuiala), suma_noua, get_data(cheltuiala), get_tip(cheltuiala))) 
         else:
             result.append(cheltuiala)
